postingAgent/requirements.txt
playwright==1.39.0
crewai==0.11.0
python-dotenv==1.0.0
openai>=1.7.1
rich==13.7.0
schedule==1.2.0